Move competency relation graph to its own component
