--- src/main/webapp/app/course/competencies/competency-management/competency-management.component.scss
@@ -1,60 +1,0 @@
-.btn-circle {
-    width: 50px !important;
-    height: 50px !important;
-    text-align: center;
-    padding: 6px 0;
-    font-size: 12px;
-    line-height: 1.428571429;
-    border-radius: 100px;
-}
-
-.btn-circle:hover {
-    box-shadow: rgba(100, 100, 111, 0.2) 0 7px 29px 0;
-}
-
-.grow {
-    transition: all 0.2s ease-in-out;
-}
-
-.grow:hover {
-    transform: scale(1.1);
-}
-
-.center-link {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.border-danger {
-    border-color: var(--danger);
-}
-
-.node {
-    text {
-        fill: var(--body-color);
-    }
-
-    rect {
-        fill: var(--primary);
-    }
-}
-
-#arrow {
-    stroke: var(--body-color);
-    fill: var(--body-color);
-}
-
-.edge {
-    stroke: var(--body-color) !important;
-    marker-end: url(#arrow);
-}
-
-.text-path {
-    fill: var(--body-color);
-}
-
-::ng-deep .accordion-body {
-    overflow: hidden;
-    max-height: 60vh;
-}
