+++ src/main/webapp/app/course/competencies/competency-management/competency-relation-graph.component.html
@@ -1,0 +1,119 @@
+<div class="container mx-0 mt-3 mb-3 mw-100">
+    <div ngbAccordion [closeOthers]="true">
+        <div ngbAccordionItem>
+            <div ngbAccordionHeader>
+                <button ngbAccordionButton>
+                    {{ 'artemisApp.competency.relation.competencyRelations' | artemisTranslate }}<span class="badge rounded-pill text-bg-warning ms-1">BETA</span>
+                </button>
+            </div>
+            <div ngbAccordionCollapse>
+                <div ngbAccordionBody>
+                    <ng-template>
+                        <form class="row p-3 g-3 align-items-center">
+                            <div class="col">
+                                <div class="form-group">
+                                    <label for="tail">{{ 'artemisApp.competency.relation.tailCompetency' | artemisTranslate }}</label>
+                                    <select
+                                        [ngClass]="{ 'border-danger': relationError }"
+                                        class="form-select"
+                                        id="tail"
+                                        name="tail"
+                                        [(ngModel)]="tailCompetencyId"
+                                        (change)="validate()"
+                                    >
+                                        @for (competency of competencies(); track competency) {
+                                            <option [value]="competency.id">{{ competency.title }}</option>
+                                        }
+                                    </select>
+                                </div>
+                            </div>
+                            <div class="col">
+                                <div class="form-group">
+                                    <label for="type">{{ 'artemisApp.competency.relation.relationType' | artemisTranslate }}</label>
+                                    <select
+                                        [ngClass]="{ 'border-danger': relationError }"
+                                        class="form-select"
+                                        id="type"
+                                        name="type"
+                                        [(ngModel)]="relationType"
+                                        (change)="validate()"
+                                    >
+                                        >
+                                        <option value="RELATES" selected>{{ 'artemisApp.competency.relation.relates' | artemisTranslate }}</option>
+                                        <option value="ASSUMES">{{ 'artemisApp.competency.relation.assumes' | artemisTranslate }}</option>
+                                        <option value="EXTENDS">{{ 'artemisApp.competency.relation.extends' | artemisTranslate }}</option>
+                                        <option value="MATCHES">{{ 'artemisApp.competency.relation.matches' | artemisTranslate }}</option>
+                                    </select>
+                                </div>
+                            </div>
+                            <div class="col">
+                                <div class="form-group">
+                                    <label for="head">{{ 'artemisApp.competency.relation.headCompetency' | artemisTranslate }}</label>
+                                    <select
+                                        [ngClass]="{ 'border-danger': relationError }"
+                                        class="form-select"
+                                        id="head"
+                                        name="head"
+                                        [(ngModel)]="headCompetencyId"
+                                        (ngModelChange)="validate()"
+                                    >
+                                        @for (competency of competencies(); track competency) {
+                                            <option [value]="competency.id">{{ competency.title }}</option>
+                                        }
+                                    </select>
+                                </div>
+                            </div>
+                            <div class="col-auto">
+                                <input
+                                    type="button"
+                                    class="btn btn-primary"
+                                    value="{{ 'artemisApp.competency.relation.createRelation' | artemisTranslate }}"
+                                    (click)="createRelation()"
+                                    [disabled]="!headCompetencyId || !tailCompetencyId || !relationType || relationError"
+                                />
+                                @if (relationError) {
+                                    <span jhiTranslate="{{ errorMessage[relationError] }}" id="relation-not-valid-text" class="invalid-feedback">
+                                        You can not create this relation.
+                                    </span>
+                                }
+                            </div>
+                        </form>
+                        <ngx-graph
+                            class="m-1 chart-container"
+                            layout="dagreCluster"
+                            [enableZoom]="false"
+                            [draggingEnabled]="false"
+                            [nodes]="nodes()"
+                            [links]="edges()"
+                            [update$]="update$"
+                        >
+                            <ng-template #defsTemplate>
+                                <svg:marker id="arrow" viewBox="0 -5 10 10" refX="8" refY="0" markerWidth="4" markerHeight="4" orient="auto">
+                                    <svg:path d="M0,-5L10,0L0,5" class="arrow-head" />
+                                </svg:marker>
+                            </ng-template>
+                            <ng-template #nodeTemplate let-node>
+                                <svg:g class="node">
+                                    <svg:rect [attr.width]="node.dimension.width" [attr.height]="node.dimension.height" />
+                                    <svg:text alignment-baseline="central" [attr.x]="10" [attr.y]="node.dimension.height / 2">
+                                        {{ node.label }}
+                                    </svg:text>
+                                </svg:g>
+                            </ng-template>
+                            <ng-template #linkTemplate let-link>
+                                <svg:g class="edge" (click)="removeRelation(link)">
+                                    <svg:path class="line" stroke-width="2" marker-end="url(#arrow)" />
+                                    <svg:text class="edge-label" text-anchor="middle">
+                                        <textPath class="text-path" [attr.href]="'#' + link.id" [style.dominant-baseline]="link.dominantBaseline" startOffset="50%">
+                                            {{ ('artemisApp.competency.relation.' + link.label.toLowerCase() | artemisTranslate).toUpperCase() }}
+                                        </textPath>
+                                    </svg:text>
+                                </svg:g>
+                            </ng-template>
+                        </ngx-graph>
+                    </ng-template>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
