+++ src/main/webapp/app/course/competencies/competency-management/competency-relation-graph.component.scss
@@ -1,0 +1,28 @@
+.accordion-body {
+    overflow: hidden;
+    max-height: 60vh;
+}
+
+.node {
+    text {
+        fill: var(--body-color);
+    }
+
+    rect {
+        fill: var(--primary);
+    }
+}
+
+.edge {
+    stroke: var(--body-color) !important;
+    marker-end: url(#arrow);
+}
+
+#arrow {
+    stroke: var(--body-color);
+    fill: var(--body-color);
+}
+
+.text-path {
+    fill: var(--body-color);
+}
