+++ src/main/webapp/app/entities/competency.model.ts
@@ -22,7 +22,6 @@ export enum CompetencyRelationError {
     CIRCULAR = 'CIRCULAR',
     SELF = 'SELF',
     EXISTING = 'EXISTING',
-    NONE = 'NONE',
 }
 
 export enum CompetencyValidators {
